+++ react/components/EXPERIMENTAL_Table/index.tsx
@@ -20,6 +20,8 @@ const Table: FC<TableProps> & TableComposites = ({
   emptyState,
   empty,
   checkboxes,
+  rowKey,
+  cellKey,
   ...props
 }) => {
   if (!measures) {
@@ -49,6 +51,8 @@ const Table: FC<TableProps> & TableComposites = ({
         {!empty && !loading && (
           <tbody>
             <Rows
+              rowKey={rowKey}
+              cellKey={cellKey}
               checkboxes={checkboxes}
               selectedDensity={selectedDensity}
               columns={columns}
@@ -75,6 +79,8 @@ export const tablePropTypes = {
   children: PropTypes.node,
   checkboxes: PropTypes.any,
   measures: PropTypes.shape(measuresPropTypes),
+  rowKey: PropTypes.func,
+  cellKey: PropTypes.func,
   containerHeight: PropTypes.number,
   empty: PropTypes.bool,
   columns: PropTypes.arrayOf(
@@ -116,7 +122,11 @@ export type TableComposites = {
   Bulk?: FC
 }
 
+<<<<<<< HEAD
 export type Items = Array<Record<string, any>>
+=======
+export interface Items extends Array<Object> { }
+>>>>>>> Rendering problems due to dynamic key generation on EXPERIMENTAL_Table.
 
 export type CellData = {
   cellData: unknown
@@ -137,5 +147,10 @@ Table.FilterBar = FilterBar
 Table.Pagination = Pagination
 Table.propTypes = tablePropTypes
 Table.Bulk = BulkActions
+
+Table.defaultProps = {
+  rowKey: ({ rowData }) => `row-${rowData.id}`,
+  cellKey: ({ cellData }) => `cell-${cellData}`,
+}
 
 export default Table